Add a keyboard-controllable Gizmo and insert it into the arena
This commit introduces an actual element that can be
keyboard-controlled around the screen. The arena defines the bounds
within which the gizmo can move around. It also controls the gizmo's
thinker logic. Gizmos are rendered on screen via a separate renderer
implementation.

There is no ECS implementation here because this primitive arena
solution should be sufficient for this app.