Fix: thumbsticks canceling each other incorrectly
Events that would arrive for one thumbstick would cause the gizmo to
move. Then events for the other thumbstick arrived, which could reset
gizmo's movement even if the original thumbstick was still held.

To prevent that issue and still support use of multiple thumbsticks, sum
the values of all thumbsticks for the same movement axis when an event
arrives for one of them.